requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
plotly>=5.18.0
pandas>=2.0.0
fpdf2>=2.7.0
Pillow>=10.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
